+++ platform/Codat/Platform/SDKConfig.cs
@@ -45,7 +45,7 @@ namespace Codat.Platform
             Client = client ?? new SpeakeasyHttpClient();
             ServerUrl = "";
             ServerIndex = 0;
-            UserAgent = "speakeasy-sdk/csharp 6.2.0 2.822.1 3.0.0 Codat.Platform";
+            UserAgent = "speakeasy-sdk/csharp 6.2.1 2.881.4 3.0.0 Codat.Platform";
             SecuritySource = null;
             Hooks = new SDKHooks();
             RetryConfig = null;