+++ platform/Codat/Platform/Utils/Utilities.cs
@@ -130,11 +130,26 @@ namespace Codat.Platform.Utils
                 && o.GetType().GetGenericTypeDefinition().IsAssignableFrom(typeof(List<>));
         }
 
+        public static bool IsModelNamespace(string ns)
+        {
+            var modelNamespaces = new[]
+            {
+                "Codat.Platform.Models.Requests",
+                "Codat.Platform.Models.Components",
+                "Codat.Platform.Models.Errors",
+                "Codat.Platform.Models.Webhooks",
+            };
+
+            return modelNamespaces.Contains(ns);
+        }
+
         public static bool IsClass(object? o)
         {
             if (o == null)
                 return false;
-            return o.GetType().IsClass && (o.GetType().FullName ?? "").StartsWith("Codat.Platform.Models");
+            if (!o.GetType().IsClass)
+                return false;
+            return IsModelNamespace(o.GetType().Namespace ?? "");
         }
 
         // TODO: code review polyfilled for IsAssignableTo