+++ platform/RELEASES.md
@@ -227,3 +227,13 @@ Based on:
 - [csharp v6.2.0] platform
 ### Releases
 - [NuGet v6.2.0] https://www.nuget.org/packages/Codat.Platform/6.2.0 - platform
+
+## 2026-04-23 10:54:45
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.9 (2.881.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [csharp v6.2.1] platform
+### Releases
+- [NuGet v6.2.1] https://www.nuget.org/packages/Codat.Platform/6.2.1 - platform